pyproject.toml
[tool.poetry]
name = "spakky-typer"
version = "0.1.0"
description = "Highly abstracted Framework core to use DDD & DI/IoC & AOP & Etc..."
authors = ["Spakky <sejong418@icloud.com>"]
readme = "README.md"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.dependencies]
python = ">=3.10"
spakky-core = "^1.2"
typer = { extras = ["all"], version = "^0.10.0" }

[tool.poetry.group.dev.dependencies]
pylint = "^3.0.2"
black = "^23.11.0"
isort = "^5.12.0"
pytest = "^7.4.3"
pytest-cov = "^4.1.0"
pytest-asyncio = "^0.21.1"
httpx = "^0.26.0"

[tool.poetry.group.ci.dependencies]
pylint = "^3.0.2"
pytest = "^7.4.3"
pytest-cov = "^4.1.0"
pytest-asyncio = "^0.21.1"

[tool.black]
line-length = 90
target-version = ["py312"]
include = '\.pyi?$'

[tool.isort]
py_version = 311
src_paths = ["spakky_typer", "tests"]
skip = [".venv", ".gitignore", ".dockerignore"]
extend_skip = [".md", ".json", ".txt"]
profile = "black"
combine_as_imports = true
remove_redundant_aliases = true
length_sort_sections = ["future", "stdlib"]

[tool.pylint]
ignore-paths = [".venv", "tests"]

[tool.pylint.'MESSAGES CONTROL']
max-line-length = 90
disable = """
    no-member,
    arguments-differ,
    invalid-field-call,
    too-many-arguments,
    missing-module-docstring,
    missing-class-docstring,
    missing-function-docstring,
    too-many-instance-attributes,
    too-few-public-methods,
    unnecessary-ellipsis,
    redefined-builtin,
    too-many-locals,
    multiple-statements,
    consider-using-f-string,
    duplicate-code,
"""

[tool.pytest.ini_options]
pythonpath = "spakky"
testpaths = "tests"
python_files = ["test_*.py"]
asyncio_mode = "auto"
addopts = """
    --cov
    --cov-report=term
    --cov-report=xml
    --no-cov-on-fail
    --strict-markers
    -p no:warnings
    -vv
"""

[tool.coverage.run]
omit = ["tests/*"]
traceback = "short"
branch = true

[tool.coverage.report]
show_missing = true
precision = 2
fail_under = 90
skip_empty = true
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "@(abc\\.)?abstractmethod",
    "@(typing\\.)?overload",
]
